+++ main.go
@@ -67,8 +67,9 @@ func (dj *mumbledj) OnDisconnect(e *gumble.DisconnectEvent) {
 // OnTextMessage event. Checks for command prefix, and calls parseCommand if it exists. Ignores
 // the incoming message otherwise.
 func (dj *mumbledj) OnTextMessage(e *gumble.TextMessageEvent) {
-	if e.Message[0] == dj.conf.General.CommandPrefix[0] && e.Message != dj.conf.General.CommandPrefix {
-		parseCommand(e.Sender, e.Sender.Name(), e.Message[1:])
+	plainMessage := gumbleutil.PlainText(&e.TextMessage)
+	if plainMessage[0] == dj.conf.General.CommandPrefix[0] && plainMessage != dj.conf.General.CommandPrefix {
+		parseCommand(e.Sender, e.Sender.Name(), plainMessage[1:])
 	}
 }